Refactored common delete modal code into its own component, implemented it into story removal modal
