+++ src/components/Projects/Stories/StoryRemovalModal.vue
@@ -1,76 +1,64 @@
 <template>
-  <div class="ui modal form">
-    <div class="header">Remove Story Confirmation</div>
-    <div
-      v-if="targetStory"
-      class="scrolling content">
-      <h3>Are you sure you want to remove "{{ targetStory.name }}"?</h3>
-      <div v-if="storyTasks.length > 0">
-        <p>The following tasks will be put into the backlog if they aren't associated with anything else</p>
-        <div class="ui three stackable cards">
-          <task-card
-            v-for="task in storyTasks"
-            :key="task.id"
-            :show-buttons="false"
-            :task="task"/>
+  <remove-modal :api-url="apiUrl">
+    <section slot="header">Remove Story Confirmation</section>
+    <section slot="content">
+      <div v-if="targetStory">
+        <h3>Are you sure you want to remove "{{ targetStory.name }}"?</h3>
+        <div v-if="storyTasks.length > 0">
+          <p>The following tasks will be put into the backlog if they aren't associated with anything else</p>
+          <div class="ui three stackable cards">
+            <task-card
+              v-for="task in storyTasks"
+              :key="task.id"
+              :show-buttons="false"
+              :task="task"/>
+          </div>
+        </div>
+        <div v-else>
+          <p>No associated tasks found with this story.</p>
+        </div>
+        <hr>
+        <div v-if="hasAssociations">
+          <p>The following items will have the story removed from their respective story list.</p>
+          <div v-if="associatedFeatures.length > 0">
+            <h4>Features</h4>
+            <feature-accordion-item
+              v-for="feature in associatedFeatures"
+              :key="`feature-${feature.id}`"
+              :id="`feature-${feature.id}`"
+              :feature="feature"
+              :active-accordion="activeAccordion"
+              @toggle-accordion-state="toggleAccordionState"
+              :name="`feature-${feature.id}`"
+              :stories="project.stories"
+              :tasks="project.tasks"
+              @click.native="refreshModal"
+              :show-buttons="false"
+            />
+          </div>
+          <div v-if="associatedSprints.length > 0">
+            <h4>Sprints</h4>
+            <sprint-accordion-item
+              v-for="sprint in associatedSprints"
+              :key="`sprint-${sprint.id}`"
+              :id="`sprint-${sprint.id}`"
+              :sprint="sprint"
+              :active-accordion="activeAccordion"
+              @toggle-accordion-state="toggleAccordionState"
+              :name="`sprint-${sprint.id}`"
+              :stories="project.stories"
+              :tasks="project.tasks"
+              @click.native="refreshModal"
+              :show-buttons="false"
+            />
+          </div>
+        </div>
+        <div v-else>
+          This story is not associated with any feature or sprint.
         </div>
       </div>
-      <div v-else>
-        <p>No associated tasks found with this story.</p>
-      </div>
-      <hr>
-      <div v-if="hasAssociations">
-        <p>The following items will have the story removed from their respective story list.</p>
-        <div v-if="associatedFeatures.length > 0">
-          <h4>Features</h4>
-          <feature-accordion-item
-            v-for="feature in associatedFeatures"
-            :key="`feature-${feature.id}`"
-            :id="`feature-${feature.id}`"
-            :feature="feature"
-            :active-accordion="activeAccordion"
-            @toggle-accordion-state="toggleAccordionState"
-            :name="`feature-${feature.id}`"
-            :stories="project.stories"
-            :tasks="project.tasks"
-            @click.native="refreshModal"
-            :show-buttons="false"
-          />
-        </div>
-        <div v-if="associatedSprints.length > 0">
-          <h4>Sprints</h4>
-          <sprint-accordion-item
-            v-for="sprint in associatedSprints"
-            :key="`sprint-${sprint.id}`"
-            :id="`sprint-${sprint.id}`"
-            :sprint="sprint"
-            :active-accordion="activeAccordion"
-            @toggle-accordion-state="toggleAccordionState"
-            :name="`sprint-${sprint.id}`"
-            :stories="project.stories"
-            :tasks="project.tasks"
-            @click.native="refreshModal"
-            :show-buttons="false"
-          />
-        </div>
-      </div>
-      <div v-else>
-        This story is not associated with any feature or sprint.
-      </div>
-      <div class="ui error message">
-        <div class="header">Error</div>
-        <p>An error has occurred</p>
-      </div>
-    </div>
-    <div class="actions">
-      <button
-        type="submit"
-        class="ui red button">
-        Remove Task
-      </button>
-      <div class="ui grey cancel button">Cancel</div>
-    </div>
-  </div>
+    </section>
+  </remove-modal>
 </template>
 
 <script>
@@ -78,13 +66,15 @@ import { mapGetters } from 'vuex'
 import SingleTaskCard from '@/components/Projects/Tasks/SingleTaskCard'
 import FeatureAccordionItem from '@/components/Projects/Features/FeatureAccordionItem'
 import SprintAccordionItem from '@/components/Projects/Sprints/SprintAccordionItem'
+import RemoveModal from '@/components/RemoveModal'
 
 /* global $ */
 export default {
   components: {
     'task-card': SingleTaskCard,
     'feature-accordion-item': FeatureAccordionItem,
-    'sprint-accordion-item': SprintAccordionItem
+    'sprint-accordion-item': SprintAccordionItem,
+    'remove-modal': RemoveModal
   },
   props: {
     targetStoryId: {
@@ -98,7 +88,6 @@ export default {
   },
   data () {
     return {
-      isLoading: false,
       $form: null,
       activeAccordion: '',
       activeSubAccordion: ''
@@ -126,16 +115,12 @@ export default {
 
       return this.targetStory.tasks.map(id => this.project.tasks[id])
     },
+    apiUrl () {
+      return `api/projects/${this.project.id}/stories/${this.targetStoryId}?member_id=${this.currentUser.id}`
+    },
     ...mapGetters(['currentUser'])
   },
   watch: {
-    isLoading (newValue) {
-      if (newValue) {
-        this.$form.addClass('loading')
-      } else {
-        this.$form.removeClass('loading')
-      }
-    },
     targetStoryId () {
       this.activeAccordion = ''
       this.$form.removeClass('error')
@@ -144,49 +129,8 @@ export default {
   },
   mounted () {
     this.$form = $(this.$el)
-    this.$form.submit((e) => {
-      e.preventDefault()
-      this.requestHandler()
-    })
-
-    this.$form.find('.actions .ui.red.button')
-      .on('click', (e) => {
-        e.preventDefault()
-        this.requestHandler()
-      })
   },
   methods: {
-    ...mapGetters(['server']),
-    async requestHandler () {
-      console.debug('Sending request to delete task', this.targetStory.id)
-      this.isLoading = true
-
-      try {
-        const result = await this.deleteStory(this.targetStory.id)
-        console.debug('result', result)
-        if (result === 'OK') {
-          this.$form.modal('hide')
-          this.$emit('update')
-        } else {
-          console.debug('Register failed!')
-          this.notifyError(result.responseJSON ? result.responseJSON.error : (result.statusText || result.error))
-        }
-      } catch (err) {
-        console.debug('Register failed', err)
-        const message = `${err.status}: ${err.statusText}`
-        this.notifyError(err.responseJSON ? err.responseJSON.error : (err.statusText || message))
-      }
-
-      this.isLoading = false
-    },
-    deleteStory (id) {
-      const apiUrl = `api/projects/${this.project.id}/stories/${id}?member_id=${this.currentUser.id}`
-      return this.server().deleteFromServer(apiUrl)
-    },
-    notifyError (message = 'An error occurred while trying to register') {
-      this.$form.find('.ui.message p').text(message)
-      this.$form.addClass('error')
-    },
     toggleAccordionState (field) {
       if (this.activeAccordion === field) {
         this.activeAccordion = ''
